feat: remove Earn Add duplicate codes
