+++ packages/web/src/components/common/select-price-range/SelectPriceRange.tsx
@@ -4,13 +4,13 @@ import { useTranslation } from "react-i18next";
 import IconInfo from "@components/common/icons/IconInfo";
 import IconStrokeArrowRight from "@components/common/icons/IconStrokeArrowRight";
 import Tooltip from "@components/common/tooltip/Tooltip";
+import { AddLiquidityPriceRage } from "@components/earn-add/earn-add-liquidity/EarnAddLiquidity";
 import {
   DefaultTick,
   PriceRangeStr,
   PriceRangeTooltip,
-  PriceRangeType
+  PriceRangeType,
 } from "@constants/option.constant";
-import { AddLiquidityPriceRage } from "@containers/earn-add-liquidity-container/EarnAddLiquidityContainer";
 import { SelectPool } from "@hooks/pool/use-select-pool";
 import { TokenModel } from "@models/token/token-model";
 import { isFetchedPools } from "@states/pool";
