+++ packages/web/src/containers/reposition-container/RepositionContainer.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
+import { AddLiquidityPriceRage } from "@components/earn-add/earn-add-liquidity/EarnAddLiquidity";
 import RepositionContent from "@components/reposition/reposition-content/RepositionContent";
 import { SwapFeeTierType } from "@constants/option.constant";
-import { AddLiquidityPriceRage } from "@containers/earn-add-liquidity-container/EarnAddLiquidityContainer";
 import { useRepositionHandle } from "@hooks/reposition/use-reposition-handle";
 import { useRepositionModalContainer } from "@hooks/reposition/use-reposition-position-modal";
