+++ packages/web/src/views/earn/EarnLayout.styles.ts
