+++ packages/web/src/views/earn/EarnLayout.tsx
