+++ packages/web/src/views/pool/pool-add/PoolAddLayout.tsx
@@ -1,4 +1,6 @@
 import React from "react";
+import { useTranslation } from "react-i18next";
+
 import { wrapper } from "./PoolAddLayout.styles";
 
 interface PoolAddLayoutProps {
@@ -18,12 +20,14 @@ const PoolAddLayout: React.FC<PoolAddLayoutProps> = ({
   footer,
   exchangeRateGraph,
 }) => {
+  const { t } = useTranslation();
+
   return (
     <div css={wrapper}>
       {header}
       <main className="content-wrap">
         <div className="title-container">
-          <h3 className="title">Earn</h3>
+          <h3 className="title">{t("business:pageHeader.earn")}</h3>
           <div className="breadcrumbs">{breadcrumbs}</div>
         </div>
         <div
