+++ packages/web/src/views/pool/pool-add/containers/exchange-rate-graph-container/ExchangeRateGraphContainer.tsx
@@ -166,7 +166,7 @@ const ExchangeRateGraphContainer: React.FC = () => {
         };
   }, [getGnotPath, poolData, isReversed]);
 
-  if (!poolPath) return null;
+  if (!poolData.tokenA?.path || !poolData.tokenB?.path) return null;
 
   return (
     <ExchangeRateGraph
